requirements.txt
flask
flask-login
flask-sqlalchemy
flask-jwt-extended
redis
python-dotenv
